app/backend/pyproject.toml
# Backend project definition for DynaChat.
#
# Managed with `uv` (not pip, not poetry). `pyproject.toml` is the dependency
# source of truth; `uv.lock` (committed) pins exact versions.
#
# Install everything:
#     cd app/backend && uv sync --all-extras
#
# Run the backend from the `app/` directory (not `app/backend/`) because the
# `backend.main:app` import path requires it:
#     cd app && uv --project backend run uvicorn backend.main:app --reload --port 8000
#
# The Dark Factory `dark-factory-fix-issue` command runs `uv sync --all-extras`
# inside the worktree before touching any code.

[project]
name = "dynachat-backend"
version = "0.1.0"
description = "DynaChat FastAPI backend — RAG over YouTube transcripts"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "aiosqlite",
    "asyncpg",
    "httpx",
    "python-dotenv",
    "openai",
    # `chunking` brings HybridChunker + its HF-based default tokenizer (which
    # HybridChunker imports at module load time, even if you don't use it).
    # `chunking-openai` brings tiktoken so we can use OpenAITokenizer instead,
    # matching our text-embedding-3-small pipeline without any HF round-trip.
    "docling-core[chunking,chunking-openai]",
    "numpy",
    "bcrypt",
    "pyjwt",
    "email-validator",
]

[project.optional-dependencies]
dev = [
    "ruff==0.9.2",
    "mypy==1.14.1",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
]

[build-system]
# This project is not a library — uv only needs a build backend because
# `[project]` requires one. hatchling is the uv/PEP 621 default.
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Tell hatch what the "package" is so it doesn't try to auto-discover.
# The backend IS the package; there's no src/ layout here.
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
# Scope: the backend package, excluding build/test artifacts.
extend-exclude = [
    ".venv",
    "venv",
    "data",
    "__pycache__",
    "dist",
    "build",
]

[tool.ruff.lint]
# Conservative starting set. Tighten later once the factory has baseline CI.
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line-too-long — format handles this
    "B008",  # FastAPI uses Depends() in defaults intentionally
]

[tool.ruff.lint.per-file-ignores]
# Tests can use unused args for fixtures and assert pytest-style.
"tests/**/*.py" = ["B011", "B017"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
# Start lenient per CLAUDE.md ("strict = false initially") and ratchet later.
strict = false
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = [
    "(^|/)\\.venv/",
    # Catches both top-level `venv/` and the legacy checked-in `backend/venv/`.
    "(^|/)venv/",
    "^data/",
    "^__pycache__/",
    "^dist/",
    "^build/",
]

[[tool.mypy.overrides]]
# Third-party libraries without type stubs.
module = ["docling_core.*", "docling.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
# -v gives enough signal to debug without being too noisy.
addopts = "-v --tb=short"
